Move `ratectrl` test logic out of main package
